Added all levels, binaries stuff

+++ source/misc/sidfile.h
@@ -11,7 +11,7 @@ class SidFile
 public:
     SidFile();
     QByteArray m_blob, m_stripped;
-    unsigned int m_loadAddress, m_initAddress, m_playAddress;
+    unsigned short m_loadAddress, m_initAddress, m_playAddress;
     QString m_outFile, m_fileName, m_path;
 
     void Load(QString filename, QString path);
